requirements.txt
aio-pika==9.3.0
aiohttp==3.8.6
aiohttp-jinja2==1.5.1
aiormq==6.7.7
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
binance-connector==3.5.0
Cerberus==1.3.5
certifi==2023.7.22
charset-normalizer==3.3.1
ciso8601==2.3.0
distlib==0.3.7
dnspython==2.4.2
docutils==0.20.1
filelock==3.12.4
frozenlist==1.4.0
gevent==23.9.1
greenlet==3.0.1
idna==3.4
Jinja2==3.1.2
m2r==0.3.1
MarkupSafe==2.1.3
marshmallow==3.20.1
mistune==0.8.4
motor==3.3.1
multidict==6.0.4
numpy==1.26.1
packaging==23.2
pamqp==3.2.1
pandas==2.1.2
pexpect==4.8.0
platformdirs==3.11.0
ptyprocess==0.7.0
pycryptodome==3.19.0
pymongo==4.5.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
ruamel.yaml==0.18.2
ruamel.yaml.clib==0.2.8
shellingham==1.5.4
six==1.16.0
termcolor==2.3.0
tomlkit==0.12.1
trafaret==2.1.1
typing-extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
virtualenv==20.24.6
websocket-client==1.6.4
yarl==1.9.2
yaspin==3.0.1
zope.event==5.0
zope.interface==6.1
